+++ apps/frontend/src/pages/quiz/libs/components/analyzing/libs/hooks/hooks.ts
@@ -1,0 +1,1 @@
+export { useAnimatedSpans } from "./use-animated-spans/use-animated-spans.hook.js";
