--- apps/frontend/src/pages/quiz/libs/components/analyzing/libs/hoook/hooks.ts
@@ -1,1 +1,0 @@
-export { useAnimatedSpans } from "./use-animated-spans/use-animated-spans.js";
